make expression support long

+++ libraries/AdaptiveExpressions/ExpressionFunctions.cs
@@ -1312,7 +1312,7 @@ namespace AdaptiveExpressions
 
             if (a.IsInteger() && b.IsInteger())
             {
-                return Convert.ToInt32(a) + Convert.ToInt32(b);
+                return Convert.ToInt64(a) + Convert.ToInt64(b);
             }
             else
             {
@@ -1329,7 +1329,7 @@ namespace AdaptiveExpressions
 
             if (a.IsInteger() && b.IsInteger())
             {
-                return Convert.ToInt32(a) - Convert.ToInt32(b);
+                return Convert.ToInt64(a) - Convert.ToInt64(b);
             }
             else
             {
@@ -1346,7 +1346,7 @@ namespace AdaptiveExpressions
 
             if (a.IsInteger() && b.IsInteger())
             {
-                return Convert.ToInt32(a) * Convert.ToInt32(b);
+                return Convert.ToInt64(a) * Convert.ToInt64(b);
             }
             else
             {
@@ -1363,7 +1363,7 @@ namespace AdaptiveExpressions
 
             if (a.IsInteger() && b.IsInteger())
             {
-                return Convert.ToInt32(a) % Convert.ToInt32(b);
+                return Convert.ToInt64(a) % Convert.ToInt64(b);
             }
             else
             {
@@ -1380,7 +1380,7 @@ namespace AdaptiveExpressions
 
             if (a.IsInteger() && b.IsInteger())
             {
-                return Convert.ToInt32(a) / Convert.ToInt32(b);
+                return Convert.ToInt64(a) / Convert.ToInt64(b);
             }
             else
             {
@@ -1867,7 +1867,7 @@ namespace AdaptiveExpressions
             return (result, error);
         }
 
-        private static (string, string) AddToTime(string timestamp, int interval, string timeUnit, string format)
+        private static (string, string) AddToTime(string timestamp, long interval, string timeUnit, string format)
         {
             string result = null;
             string error = null;
@@ -2767,7 +2767,7 @@ namespace AdaptiveExpressions
                         {
                             object value = null;
                             string error;
-                            if (Convert.ToInt32(args[1]) == 0)
+                            if (Convert.ToInt64(args[1]) == 0)
                             {
                                 error = $"Cannot mod by 0";
                             }
@@ -2831,7 +2831,7 @@ namespace AdaptiveExpressions
                         args =>
                         {
                             var operands = ResolveListValue(args[0]).OfType<object>().ToList();
-                            return operands.All(u => u.IsInteger()) ? operands.Sum(u => Convert.ToInt32(u)) : operands.Sum(u => Convert.ToSingle(u));
+                            return operands.All(u => u.IsInteger()) ? operands.Sum(u => Convert.ToInt64(u)) : operands.Sum(u => Convert.ToSingle(u));
                         },
                         VerifyNumericList),
                     ReturnType.Number,
@@ -3390,7 +3390,7 @@ namespace AdaptiveExpressions
                             {
                                 var format = (args.Count() == 4) ? (string)args[3] : DefaultDateTimeFormat;
                                 Func<DateTime, DateTime> timeConverter;
-                                (timeConverter, error) = DateTimeConverter(Convert.ToInt32(args[1]), string2);
+                                (timeConverter, error) = DateTimeConverter(Convert.ToInt64(args[1]), string2);
                                 if (error == null)
                                 {
                                     (value, error) = ParseISOTimestamp(string0, dt => timeConverter(dt).ToString(format));
@@ -3487,7 +3487,7 @@ namespace AdaptiveExpressions
                             {
                                 var format = (args.Count() == 3) ? (string)args[2] : DefaultDateTimeFormat;
                                 Func<DateTime, DateTime> timeConverter;
-                                (timeConverter, error) = DateTimeConverter(Convert.ToInt32(args[0]), string1, false);
+                                (timeConverter, error) = DateTimeConverter(Convert.ToInt64(args[0]), string1, false);
                                 if (error == null)
                                 {
                                     value = timeConverter(DateTime.Now).ToString(format);
@@ -3517,7 +3517,7 @@ namespace AdaptiveExpressions
                             {
                                 var format = (args.Count() == 3) ? (string)args[2] : DefaultDateTimeFormat;
                                 Func<DateTime, DateTime> timeConverter;
-                                (timeConverter, error) = DateTimeConverter(Convert.ToInt32(args[0]), string1);
+                                (timeConverter, error) = DateTimeConverter(Convert.ToInt64(args[0]), string1);
                                 if (error == null)
                                 {
                                     value = timeConverter(DateTime.Now).ToString(format);
@@ -3596,7 +3596,7 @@ namespace AdaptiveExpressions
                             var format = (args.Count() == 4) ? (string)args[3] : DefaultDateTimeFormat;
                             if (args[0] is string timestamp && args[1].IsInteger() && args[2] is string timeUnit)
                             {
-                                (value, error) = AddToTime(timestamp, Convert.ToInt32(args[1]), timeUnit, format);
+                                (value, error) = AddToTime(timestamp, Convert.ToInt64(args[1]), timeUnit, format);
                             }
                             else
                             {
@@ -3856,7 +3856,7 @@ namespace AdaptiveExpressions
 
                 // Conversions
                 new ExpressionEvaluator(ExpressionType.Float, Apply(args => CultureInvariantDoubleConvert(args[0])), ReturnType.Number, ValidateUnary),
-                new ExpressionEvaluator(ExpressionType.Int, Apply(args => Convert.ToInt32(args[0])), ReturnType.Number, ValidateUnary),
+                new ExpressionEvaluator(ExpressionType.Int, Apply(args => Convert.ToInt64(args[0])), ReturnType.Number, ValidateUnary),
                 new ExpressionEvaluator(ExpressionType.Binary, Apply(args => ToBinary(args[0].ToString()), VerifyString), ReturnType.String, ValidateUnary),
                 new ExpressionEvaluator(ExpressionType.Base64, Apply(args => Convert.ToBase64String(System.Text.Encoding.UTF8.GetBytes(args[0].ToString())), VerifyString), ReturnType.String, ValidateUnary),
                 new ExpressionEvaluator(ExpressionType.Base64ToBinary, Apply(args => ToBinary(args[0].ToString()), VerifyString), ReturnType.String, ValidateUnary),
